requirements.txt
# Core
fastapi==0.125.0
uvicorn==0.38.0
python-multipart==0.0.21

# Database (PostgreSQL)
SQLAlchemy==2.0.45
psycopg2-binary>=2.9.9

# Auth
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt>=4.0.0

# Config & validation
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
email-validator>=2.0.0

# HTTP
httpx>=0.27.0
starlette==0.50.0

# Optional / transitive
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
pyasn1==0.6.1
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
